Removed Python2 script
